Rename page_input_providers to just providers
