+++ example/example/autoextract.py
@@ -9,7 +9,7 @@ from twisted.internet.defer import inlineCallbacks, returnValue
 from twisted.internet.threads import deferToThread
 
 from scrapy_poet.repository import provides
-from scrapy_poet.page_input_providers import PageObjectInputProvider
+from scrapy_poet.providers import PageObjectInputProvider
 from web_poet.pages import ItemPage
 from scrapy_poet.utils import DummyResponse
