+++ scrapy_poet/utils.py
@@ -12,7 +12,7 @@ from twisted.internet.defer import inlineCallbacks, returnValue
 
 from web_poet.pages import ItemPage, is_injectable
 from scrapy_poet.repository import providers
-from scrapy_poet.page_input_providers import PageObjectInputProvider
+from scrapy_poet.providers import PageObjectInputProvider
 
 _CALLBACK_FOR_MARKER = '__scrapy_poet_callback'
